requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Firebase
firebase-admin==6.4.0

# LiveKit
livekit-api==1.0.3

# OpenAI
openai==1.3.0

# Database & Cache
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# WebSocket
websockets==12.0

# Environment & Config
python-dotenv==1.0.0

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging
structlog==23.2.0

# Async support
aiofiles==23.2.1

# HTTP Client
requests==2.31.0 